lib: Always build modules as shared libs
Using "target_link_libraries" with OBJECT libraries in cmake appears to
be impossible for most cmake versions (my Fedora 30 setup seems to allow
it though?), so we restrict the build to solibs for now.

As part of this change, the OpenCV dependencies of the modules are made
part of their public link interfaces (through target_link_libraries).

The seraphim_messages target (which is the compiled Protobuf 3 classes)
is compiled as position independent code (PIC) to allow shared libs to
link it in.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>