lib: face: recognizer: Perform grayscale conversion if required
The internal m_faces buffer will still hold the multi-channel (i.e. BGR)
images. We are feeding grayscale images to the algorithms exclusively to
conform to the OpenCV algo API constraints. This enables us not having to
worry about the colorspace conversion on the caller side of things.

Only the predict() method handles UMat and Mat instances in seperate
codepaths for T-API acceleration benefits as the other two, "train" and
"update" only involve small input images and are unlikely to benefit much
from T-API calls.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>