apps: backend: Use custom Image2D type for IPC
Prior to this change, clients were forced to link against OpenCV to be
able to provide images in a format the backend was able to understand.
With this change, we break that dependency and allow clients to pass
buffers from arbitrary sources, provided they tell us what format it has.

The backend has a new utils collection to facilitate decoding the received
image buffer into an OpenCV compatible matrix type. At the moment, only
raw (YUYV) and motion jpeg (MJPG) encoded frames are supported.
Support for other types will be added as needed.

As part of this change, the static libs are linked into the backend in
PUBLIC interface mode now so depedencies are propagated properly.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>