apps: frontend: v4l2: Actually use multiple device buffers
In the mmap case, device memory is mapped into userspace.
We request which device buffer to fill by setting the v4l2_buf.index
field to the index of the buffer.
This is needed for both VIDIOC_QBUF and VIDIOC_DQBUF.

This way, clients may still hold (valid) references to buffer data even if
the current buffer index has changed in case of multiple device buffers
being used.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>