apps: frontend: QImageProvider: Fix double free bug
By default, a QImage instance will attempt to free its internal pixel
buffer. However, when doing a shallow copy of QImage, this is not what
you want since it would actually attempt to free the data twice: once
in the destructor of the shallow clone and once in the destructor of the
originating instance (which owns the pixel data).

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>