lib: core: Rework image infrastructure
Define an abstract image class that may be implemented to support
alternative buffers (such as OpenCL contexts, etc).
In the future, the interfaces (face detector, etc) will be changed to
accept this abstract image class and operate on that.

However, for that to be possible, there's quite a lot more work ahead,
including the need for a new abstract buffer class for holding data,
maybe a two dimensional array to act as image back buffer.

The new abstract back buffer implementation shall then be guaranteed
to be supported by each IImage instance.

A quick summary of other high-level changes:

 * move qt image utils out of the library
 * opencv image utils stay in the core lib for now
 * make image buffer an instance member
 * images must at least support RGB(A) pixel values, no matter what
   their underlying buffer format is

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>