lib: object: dnn: Fix DetectionOutput box scaling
With networks that produce "DetectionOutput" output layers (such as
MobileNet v2), left, top, right, bottom are usually floating point
values in the [0, 1] range. That means they need to be scaled by the
image width/height.

The code to do so was already in place, but not quite correct: if
just one of the values got clamped to 1 as part of the integer casting,
there was a good change of either width or height ending up with a
value of 2, thus width * height would be 2 in most cases and no scaling
would happen.

In the extreme case, both width and height would end up being 2, so
improve the scaling condition by checking for values lower than 5.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>