ci: Enable Docker caching
Combine the relevant images (fedora & builder) into one
archive so we can save much of the space it would take
to save each of them individually.
This is possible because fedora is the base image of
builder, thus most of their contents are shared.

The overall caching mechanism here is quite complex:

 - cache all docker images required for our build
   (builder tag and all base images, i.e. fedora)

 - if the hash of the image is different from the one
   in the cache, write the new image to the cache

Based on the following hint I found on GitHub:
https://github.com/travis-ci/travis-ci/issues/5358#issuecomment-248915326

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>