lib: Add new GUI module
A reference OpenGL window implementation is included, using GLFW
and GLEW. Most pixelformats are supported as well (except for Y16
for now, because that somehow does not work).

A basic sample demonstrating the GL window usage is included as
well.

In the future, a Vulkan window implementation will be added
using GLFW as well.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>