lib: Make all modules use the interoperability module
Almost all of our implementation modules depend on OpenCV
to carry out the actual computations. The task of converting
between sph::core::Image and cv::Mat is now deferred to the
streamlined iop module.

This allows us to drop the OpenCV dependency from the core
module, which is a nice benefit.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>