lib: ipc: semaphore: Fix design defects
Creating semaphores assumes they already reside in a
shared memory segment, thus we should just take the
address of the existing handle and pass it on to
sem_init (sem_open on Darwin).

Do not create the handle on the stack as class member
in any case, but use the handle returned by POSIX
functions.

Remove the unneeded ptr() member function to avoid
misguided use of it (should not be required for normal
semaphore operations).

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>