lib: ipc: shm: Initial concurrent reader/writer support
Introduce a concept of message source/destination.
Based on these fields, a client may determine whether or
not it is legal to read/write a message.

The status field is no longer needed, thus removed.
Instead, there is a new type field to so clients don't
overwrite their own request/response messages.

Still on the TODO list is a fair scheduling queue, that
prevents reader/writer starvation by queuing their requests
for execution in the order they came in (FIFO).

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>