ci: Only keep one docker image cached
"docker load -i" can only handle one archive at a time
and it does not make sense to keep more than one around
anyway.

In case of a new archive, the docker images/layers obviously
changed, so we can just discard the old archive and remove
it from the cache.

Implement a safeguard for the future: whenever there are
multiple archives in the cache, something went wrong during
cache save, so remove all of them and start fresh.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>