lib: dnn: Add support for Vulkan compute target
The Vulkan target requires the VKCOM backend, so we now
query the available backends for their supported targets
and configure the underlying DNN impl accordingly.

Unfortunately we can only query available backends starting
from OpenCV 3.4.5, so for now just try to force the issue.
Vulkan support is new as of OpenCV 4.0.0.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>