lib: ipc: net: Rework socket architecture
Use C++ exceptions to propagate (exceptional) errors up the stack.
Remove the (deprecated) TCPStream class. It was originally designed
to work as a segment buffer with Nagles' algorithm being disabled.
That never came out as being necessary in the end though.

In the future, when other socket types and protocols (e.g. UDP) are
supported, a packet buffer class to build up bursts (to avoid network
congestion) may be introduced again.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>