lib: face: Directly convert from sph to dlib
Remove the intermediate OpenCV matrix type usage.
The current code is still not optimal because it always copies the pixel
buffer - even in the single channel input case where we could just use
the existing buffer directly. That requires more boilerplate code
though, which I plan to implement in the iop library.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>