lib: core: matrix: Add BLAS gemm support
BLAS works on single precision (float) and double precision (double)
values so the input and output matrices have to be converted.
For that reason, the CoreMatrix<T> copy constructor now takes matrices
of any type and does the conversion in-place if necessary.

A matrix sample has been added to measure the performance.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>